Related work items: #18054, #18445

+++ src/CPG.Application/UseCases/Users/ViewModel/UserViewModel.cs
@@ -4,7 +4,7 @@ public class UserViewModel
 {
     public string IDPId { get; set; }
     public string NationalCode { get; set; }
-    public int CompanyId { get; set; }
+    public long CompanyId { get; set; }
     public string FirstName { get; set; }
     public string LastName { get; set; }
     public string PhoneNumber { get; set; }
